+++ compiler/src/main/java/solar/blaz/rondel/compiler/model/ComponentModel.java
@@ -32,6 +32,7 @@ public class ComponentModel {
     public TypeElement[] components;
     public Element element;
     public TypeMirror parent;
+    public TypeElement scope;
 
     public InjectorModel injector;