+++ compiler/src/test/java/solar/blaz/rondel/compiler/ActivityInjectorTest.java
@@ -19,7 +19,7 @@ package solar.blaz.rondel.compiler;
 import com.google.common.collect.ImmutableList;
 import com.google.testing.compile.JavaFileObjects;
 
-import org.junit.gen5.api.Test;
+import org.junit.Test;
 
 import javax.tools.JavaFileObject;
 
@@ -94,25 +94,25 @@ public class ActivityInjectorTest {
                 "    \n" +
                 "}");
 
-        JavaFileObject expectedComponent = JavaFileObjects.forSourceString("test.ui.RondelTestActivityComponent", "package test.ui;\n" +
-                "\n" +
-                "import dagger.Subcomponent;\n" +
-                "import solar.blaz.rondel.BaseComponent;\n" +
-                "import solar.blaz.rondel.ViewScope;\n" +
-                "\n" +
-                "@Subcomponent(\n" +
-                "        modules = { TestModule.class }\n" +
-                ")\n" +
-                "@ViewScope\n" +
-                "public interface RondelTestActivityComponent extends BaseComponent {\n" +
-                "    void inject(TestActivity view);\n" +
-                "    \n" +
-                "    @Subcomponent.Builder\n" +
-                "    interface Builder {\n" +
-                "        Builder testModule(TestModule module);\n" +
-                "        RondelTestActivityComponent build();\n" +
-                "    }\n" +
-                "}");
+        JavaFileObject expectedComponent = JavaFileObjects.forSourceString("test.ui.RondelTestActivityComponent", "package test.ui;\n"
+                + "\n"
+                + "import dagger.Subcomponent;\n"
+                + "import solar.blaz.rondel.ActivityScope;\n"
+                + "import solar.blaz.rondel.BaseComponent;\n"
+                + "\n"
+                + "@Subcomponent(\n"
+                + "        modules = { TestModule.class }\n"
+                + ")\n"
+                + "@ActivityScope\n"
+                + "public interface RondelTestActivityComponent extends BaseComponent {\n"
+                + "    void inject(TestActivity view);\n"
+                + "    \n"
+                + "    @Subcomponent.Builder\n"
+                + "    interface Builder {\n"
+                + "        Builder testModule(TestModule module);\n"
+                + "        RondelTestActivityComponent build();\n"
+                + "    }\n"
+                + "}");
 
         assertAbout(javaSources())
                 .that(ImmutableList.of(appFile, moduleFile, componentFile, activityFile, activityModuleFile))
@@ -215,24 +215,24 @@ public class ActivityInjectorTest {
                 "    \n" +
                 "}");
 
-        JavaFileObject expectedComponent = JavaFileObjects.forSourceString("test.ui.RondelTestActivityComponent", "package test.ui;\n" +
-                "\n" +
-                "import dagger.Subcomponent;\n" +
-                "import solar.blaz.rondel.BaseComponent;\n" +
-                "import solar.blaz.rondel.ViewScope;\n" +
-                "\n" +
-                "@Subcomponent\n" +
-                "@ViewScope\n" +
-                "public interface RondelTestActivityComponent extends BaseComponent, TestComponent {\n" +
-                "    \n" +
-                "    void inject(TestActivity view);\n" +
-                "\n" +
-                "    @Subcomponent.Builder\n" +
-                "    interface Builder {\n" +
-                "        RondelTestActivityComponent build();\n" +
-                "    }\n" +
-                "    \n" +
-                "}");
+        JavaFileObject expectedComponent = JavaFileObjects.forSourceString("test.ui.RondelTestActivityComponent", "package test.ui;\n"
+                + "\n"
+                + "import dagger.Subcomponent;\n"
+                + "import solar.blaz.rondel.ActivityScope;\n"
+                + "import solar.blaz.rondel.BaseComponent;\n"
+                + "\n"
+                + "@Subcomponent\n"
+                + "@ActivityScope\n"
+                + "public interface RondelTestActivityComponent extends BaseComponent, TestComponent {\n"
+                + "    \n"
+                + "    void inject(TestActivity view);\n"
+                + "\n"
+                + "    @Subcomponent.Builder\n"
+                + "    interface Builder {\n"
+                + "        RondelTestActivityComponent build();\n"
+                + "    }\n"
+                + "    \n"
+                + "}");
 
         assertAbout(javaSources())
                 .that(ImmutableList.of(appFile, moduleFile, componentFile, activityFile, activityComponentFile))
@@ -499,27 +499,27 @@ public class ActivityInjectorTest {
                 "    \n" +
                 "}");
 
-        JavaFileObject expectedComponent = JavaFileObjects.forSourceString("test.ui.RondelTestActivityComponent", "package test.ui;\n" +
-                "\n" +
-                "import dagger.Subcomponent;\n" +
-                "import solar.blaz.rondel.BaseComponent;\n" +
-                "import solar.blaz.rondel.ViewScope;\n" +
-                "\n" +
-                "@Subcomponent(\n" +
-                "        modules = { TestModule.class }\n" +
-                ")\n" +
-                "@ViewScope\n" +
-                "public interface RondelTestActivityComponent extends BaseComponent, TestComponent {\n" +
-                "    \n" +
-                "    void inject(TestActivity view);\n" +
-                "\n" +
-                "    @Subcomponent.Builder\n" +
-                "    interface Builder {\n" +
-                "        Builder testModule(TestModule module);\n" +
-                "        RondelTestActivityComponent build();\n" +
-                "    }\n" +
-                "    \n" +
-                "}");
+        JavaFileObject expectedComponent = JavaFileObjects.forSourceString("test.ui.RondelTestActivityComponent", "package test.ui;\n"
+                + "\n"
+                + "import dagger.Subcomponent;\n"
+                + "import solar.blaz.rondel.ActivityScope;\n"
+                + "import solar.blaz.rondel.BaseComponent;\n"
+                + "\n"
+                + "@Subcomponent(\n"
+                + "        modules = { TestModule.class }\n"
+                + ")\n"
+                + "@ActivityScope\n"
+                + "public interface RondelTestActivityComponent extends BaseComponent, TestComponent {\n"
+                + "    \n"
+                + "    void inject(TestActivity view);\n"
+                + "\n"
+                + "    @Subcomponent.Builder\n"
+                + "    interface Builder {\n"
+                + "        Builder testModule(TestModule module);\n"
+                + "        RondelTestActivityComponent build();\n"
+                + "    }\n"
+                + "    \n"
+                + "}");
 
         assertAbout(javaSources())
                 .that(ImmutableList.of(appFile, moduleFile, componentFile, activityFile, activityModuleFile, activityComponentFile))
@@ -645,27 +645,27 @@ public class ActivityInjectorTest {
                 "    \n" +
                 "}");
 
-        JavaFileObject expectedComponent = JavaFileObjects.forSourceString("test.ui.RondelTestActivityComponent", "package test.ui;\n" +
-                "\n" +
-                "import dagger.Subcomponent;\n" +
-                "import solar.blaz.rondel.BaseComponent;\n" +
-                "import solar.blaz.rondel.ViewScope;\n" +
-                "\n" +
-                "@Subcomponent(\n" +
-                "        modules = { TestModule.class }\n" +
-                ")\n" +
-                "@ViewScope\n" +
-                "public interface RondelTestActivityComponent extends BaseComponent, TestComponent {\n" +
-                "    \n" +
-                "    void inject(TestActivity view);\n" +
-                "    \n" +
-                "    @Subcomponent.Builder\n" +
-                "    interface Builder {\n" +
-                "        Builder testModule(TestModule module);\n" +
-                "        RondelTestActivityComponent build();\n" +
-                "    }\n" +
-                "    \n" +
-                "}");
+        JavaFileObject expectedComponent = JavaFileObjects.forSourceString("test.ui.RondelTestActivityComponent", "package test.ui;\n"
+                + "\n"
+                + "import dagger.Subcomponent;\n"
+                + "import solar.blaz.rondel.ActivityScope;\n"
+                + "import solar.blaz.rondel.BaseComponent;\n"
+                + "\n"
+                + "@Subcomponent(\n"
+                + "        modules = { TestModule.class }\n"
+                + ")\n"
+                + "@ActivityScope\n"
+                + "public interface RondelTestActivityComponent extends BaseComponent, TestComponent {\n"
+                + "    \n"
+                + "    void inject(TestActivity view);\n"
+                + "    \n"
+                + "    @Subcomponent.Builder\n"
+                + "    interface Builder {\n"
+                + "        Builder testModule(TestModule module);\n"
+                + "        RondelTestActivityComponent build();\n"
+                + "    }\n"
+                + "    \n"
+                + "}");
 
         assertAbout(javaSources())
                 .that(ImmutableList.of(appFile, moduleFile, componentFile, activityFile, activityModuleFile, activityComponentFile))
@@ -870,15 +870,15 @@ public class ActivityInjectorTest {
         JavaFileObject expectedComponent = JavaFileObjects.forSourceString("test.ui.activity.RondelTestActivityComponent", "package test.ui.activity;\n" +
                 "\n" +
                 "import dagger.Subcomponent;\n" +
+                "import solar.blaz.rondel.ActivityScope;\n" +
                 "import solar.blaz.rondel.BaseComponent;\n" +
-                "import solar.blaz.rondel.ViewScope;\n" +
                 "import test.ui.component.TestComponent;\n" +
                 "import test.ui.module.TestModule;\n" +
                 "\n" +
                 "@Subcomponent(\n" +
                 "        modules = { TestModule.class }\n" +
                 ")\n" +
-                "@ViewScope\n" +
+                "@ActivityScope\n" +
                 "public interface RondelTestActivityComponent extends BaseComponent, TestComponent {\n" +
                 "    \n" +
                 "    void inject(TestActivity view);\n" +
@@ -1043,13 +1043,13 @@ public class ActivityInjectorTest {
         JavaFileObject expectedComponent = JavaFileObjects.forSourceString("test.ui.RondelTestActivityComponent", "package test.ui;\n" +
                 "\n" +
                 "import dagger.Subcomponent;\n" +
+                "import solar.blaz.rondel.ActivityScope;\n" +
                 "import solar.blaz.rondel.BaseComponent;\n" +
-                "import solar.blaz.rondel.ViewScope;\n" +
                 "\n" +
                 "@Subcomponent(\n" +
                 "        modules = { TestModule.class, TestModule2.class }\n" +
                 ")\n" +
-                "@ViewScope\n" +
+                "@ActivityScope\n" +
                 "public interface RondelTestActivityComponent extends BaseComponent, TestComponent {\n" +
                 "    \n" +
                 "    void inject(TestActivity view);\n" +
@@ -1220,13 +1220,13 @@ public class ActivityInjectorTest {
         JavaFileObject expectedComponent = JavaFileObjects.forSourceString("test.ui.RondelTestActivityComponent", "package test.ui;\n" +
                 "\n" +
                 "import dagger.Subcomponent;\n" +
+                "import solar.blaz.rondel.ActivityScope;\n" +
                 "import solar.blaz.rondel.BaseComponent;\n" +
-                "import solar.blaz.rondel.ViewScope;\n" +
                 "\n" +
                 "@Subcomponent(\n" +
                 "        modules = { TestModule.class, TestModule2.class }\n" +
                 ")\n" +
-                "@ViewScope\n" +
+                "@ActivityScope\n" +
                 "public interface RondelTestActivityComponent extends BaseComponent, TestComponent {\n" +
                 "    \n" +
                 "    void inject(TestActivity view);\n" +
@@ -1246,6 +1246,255 @@ public class ActivityInjectorTest {
                 .compilesWithoutError()
                 .and()
                 .generatesSources(expectedInjector, expectedComponent);
+
+    }
+
+    @Test
+    public void testCustomActivityScope() throws Exception {
+
+        JavaFileObject scopeFile = JavaFileObjects.forSourceString("test.CustomScope", "package test;\n"
+                + "\n"
+                + "import java.lang.annotation.Retention;\n"
+                + "import java.lang.annotation.RetentionPolicy;\n"
+                + "\n"
+                + "import javax.inject.Scope;\n"
+                + "\n"
+                + "@Scope @Retention(RetentionPolicy.RUNTIME)\n"
+                + "public @interface CustomScope {\n"
+                + "}\n");
+
+        JavaFileObject moduleFile = JavaFileObjects.forSourceString("test.AppModule", "package test;\n" +
+                "\n" +
+                "import dagger.Module;\n" +
+                "\n" +
+                "@Module\n" +
+                "public class AppModule {\n" +
+                "    \n" +
+                "    public AppModule(TestApp app) {\n" +
+                "        \n" +
+                "    }\n" +
+                "    \n" +
+                "}");
+
+        JavaFileObject componentFile = JavaFileObjects.forSourceString("test.AppComponent", "package test;\n" +
+                "\n" +
+                "import dagger.Component;\n" +
+                "\n" +
+                "@Component\n" +
+                "public interface AppComponent {\n" +
+                "    \n" +
+                "}");
+
+        JavaFileObject appFile = JavaFileObjects.forSourceString("test.TestApp", "package test;\n" +
+                "\n" +
+                "import android.app.Application;\n" +
+                "import solar.blaz.rondel.App;\n" +
+                "import solar.blaz.rondel.AppComponentProvider;\n" +
+                "\n" +
+                "@App(\n" +
+                "        components = AppComponent.class,\n" +
+                "        modules = AppModule.class\n" +
+                ")\n" +
+                "public class TestApp extends Application implements AppComponentProvider {\n" +
+                "    public RondelTestAppComponent getComponent() {\n" +
+                "        return null;\n" +
+                "    }\n" +
+                "}");
+
+        JavaFileObject activityModuleFile = JavaFileObjects.forSourceString("test.ui.TestModule", "package test.ui;\n" +
+                "\n" +
+                "import dagger.Module;\n" +
+                "\n" +
+                "@Module\n" +
+                "public class TestModule {\n" +
+                "    \n" +
+                "}");
+
+        JavaFileObject activityComponentFile = JavaFileObjects.forSourceString("test.ui.TestComponent", "package test.ui;\n" +
+                "\n" +
+                "import dagger.Component;\n" +
+                "\n" +
+                "@Component\n" +
+                "public interface TestComponent {\n" +
+                "    \n" +
+                "}");
+
+        JavaFileObject activityFile = JavaFileObjects.forSourceString("test.ui.TestActivity", "package test.ui;\n"
+                + "\n"
+                + "import android.app.Activity;\n"
+                + "import solar.blaz.rondel.Rondel;\n"
+                + "import test.CustomScope;\n"
+                + "\n"
+                + "@Rondel(\n"
+                + "        components = TestComponent.class,\n"
+                + "        modules = TestModule.class,\n"
+                + "        scope = CustomScope.class\n"
+                + ")\n"
+                + "public class TestActivity extends Activity {\n"
+                + "    \n"
+                + "}");
+
+        JavaFileObject expectedInjector = JavaFileObjects.forSourceString("test.ui.RondelTestActivity", "package test.ui;\n" +
+                "\n" +
+                "import test.RondelTestAppComponent;\n" +
+                "import test.TestApp;\n" +
+                "\n" +
+                "class RondelTestActivity {\n" +
+                "    \n" +
+                "    private static RondelTestAppComponent component;\n" +
+                "    private static TestModule testModule;\n" +
+                "    \n" +
+                "    private static RondelTestAppComponent getComponent(TestApp app) {\n" +
+                "        if (component != null) {\n" +
+                "            return component;\n" +
+                "        } else {\n" +
+                "            return (RondelTestAppComponent) app.getComponent();\n" +
+                "        }\n" +
+                "    }\n" +
+                "    \n" +
+                "    public static void setComponent(RondelTestAppComponent component) {\n" +
+                "        RondelTestActivity.component = component;\n" +
+                "    }\n" +
+                "    \n" +
+                "    public static RondelTestActivityComponent inject(TestActivity injectie) {\n" +
+                "        TestApp app = (TestApp) injectie.getApplicationContext();\n" +
+                "        RondelTestActivityComponent component = getComponent(app).rondelTestActivityComponentBuilder()\n" +
+                "                .testModule(getTestModule(injectie))\n" +
+                "                .build();\n" +
+                "        component.inject(injectie);\n" +
+                "        return component;\n" +
+                "    }\n" +
+                "    \n" +
+                "    public static void setTestModule(TestModule module) {\n" +
+                "        testModule = module;\n" +
+                "    }\n" +
+                "    \n" +
+                "    private static TestModule getTestModule(TestActivity injectie) {\n" +
+                "        if (testModule != null) {\n" +
+                "            return testModule;\n" +
+                "        } else {\n" +
+                "            return new TestModule();\n" +
+                "        }\n" +
+                "    }\n" +
+                "    \n" +
+                "}");
+
+        JavaFileObject expectedComponent = JavaFileObjects.forSourceString("test.ui.RondelTestActivityComponent", "package test.ui;\n"
+                + "\n"
+                + "import dagger.Subcomponent;\n"
+                + "import solar.blaz.rondel.BaseComponent;\n"
+                + "import test.CustomScope;\n"
+                + "\n"
+                + "@Subcomponent(\n"
+                + "        modules = { TestModule.class }\n"
+                + ")\n"
+                + "@CustomScope\n"
+                + "public interface RondelTestActivityComponent extends BaseComponent, TestComponent {\n"
+                + "    \n"
+                + "    void inject(TestActivity view);\n"
+                + "\n"
+                + "    @Subcomponent.Builder\n"
+                + "    interface Builder {\n"
+                + "        Builder testModule(TestModule module);\n"
+                + "        RondelTestActivityComponent build();\n"
+                + "    }\n"
+                + "    \n"
+                + "}");
+
+        assertAbout(javaSources())
+                .that(ImmutableList.of(scopeFile, appFile, moduleFile, componentFile, activityFile, activityModuleFile, activityComponentFile))
+                .processedWith(new RondelProcessor(), new ComponentProcessor())
+                .compilesWithoutError()
+                .and()
+                .generatesSources(expectedInjector, expectedComponent);
+
+    }
+
+
+    @Test
+    public void testCustomActivityScopeNotInteface() throws Exception {
+
+        JavaFileObject scopeFile = JavaFileObjects.forSourceString("test.CustomScope", "package test;\n"
+                + "\n"
+                + "public class CustomScope {\n"
+                + "}\n");
+
+        JavaFileObject moduleFile = JavaFileObjects.forSourceString("test.AppModule", "package test;\n" +
+                "\n" +
+                "import dagger.Module;\n" +
+                "\n" +
+                "@Module\n" +
+                "public class AppModule {\n" +
+                "    \n" +
+                "    public AppModule(TestApp app) {\n" +
+                "        \n" +
+                "    }\n" +
+                "    \n" +
+                "}");
+
+        JavaFileObject componentFile = JavaFileObjects.forSourceString("test.AppComponent", "package test;\n" +
+                "\n" +
+                "import dagger.Component;\n" +
+                "\n" +
+                "@Component\n" +
+                "public interface AppComponent {\n" +
+                "    \n" +
+                "}");
+
+        JavaFileObject appFile = JavaFileObjects.forSourceString("test.TestApp", "package test;\n" +
+                "\n" +
+                "import android.app.Application;\n" +
+                "import solar.blaz.rondel.App;\n" +
+                "import solar.blaz.rondel.AppComponentProvider;\n" +
+                "\n" +
+                "@App(\n" +
+                "        components = AppComponent.class,\n" +
+                "        modules = AppModule.class\n" +
+                ")\n" +
+                "public class TestApp extends Application implements AppComponentProvider {\n" +
+                "    public RondelTestAppComponent getComponent() {\n" +
+                "        return null;\n" +
+                "    }\n" +
+                "}");
+
+        JavaFileObject activityModuleFile = JavaFileObjects.forSourceString("test.ui.TestModule", "package test.ui;\n" +
+                "\n" +
+                "import dagger.Module;\n" +
+                "\n" +
+                "@Module\n" +
+                "public class TestModule {\n" +
+                "    \n" +
+                "}");
+
+        JavaFileObject activityComponentFile = JavaFileObjects.forSourceString("test.ui.TestComponent", "package test.ui;\n" +
+                "\n" +
+                "import dagger.Component;\n" +
+                "\n" +
+                "@Component\n" +
+                "public interface TestComponent {\n" +
+                "    \n" +
+                "}");
+
+        JavaFileObject activityFile = JavaFileObjects.forSourceString("test.ui.TestActivity", "package test.ui;\n"
+                + "\n"
+                + "import android.app.Activity;\n"
+                + "import solar.blaz.rondel.Rondel;\n"
+                + "import test.CustomScope;\n"
+                + "\n"
+                + "@Rondel(\n"
+                + "        components = TestComponent.class,\n"
+                + "        modules = TestModule.class,\n"
+                + "        scope = CustomScope.class\n"
+                + ")\n"
+                + "public class TestActivity extends Activity {\n"
+                + "    \n"
+                + "}");
+
+        assertAbout(javaSources())
+                .that(ImmutableList.of(scopeFile, appFile, moduleFile, componentFile, activityFile, activityModuleFile, activityComponentFile))
+                .processedWith(new RondelProcessor(), new ComponentProcessor())
+                .failsToCompile()
+                .withErrorContaining("Scope has to bo an annotation");
 
     }
 }